11382 did not save; update save

+++ 22_10/01/11382.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main(void) {
-	int a, b, c;
+	unsigned long long a, b, c;
 
-	scanf("%d %d %d", &a, &b, &c);
-	printf("%d", a + b + c);
+	scanf("%lld %lld %lld", &a, &b, &c);
+	printf("%lld", a + b + c);
 }
